+++ src/main/java/staryhroft/weatherapp/service/WeatherApiResponse.java
@@ -5,11 +5,10 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 public class WeatherApiResponse {
-    private BigDecimal temperature;
+    private Double temperature;
 }
